+++ tests/Unit/Service/ImportServiceTest.php
@@ -70,8 +70,9 @@ class ImportServiceTest extends TestCase
             ->disableOriginalConstructor()
             ->getMock();
         $this->purifier = $this
-            ->getMockBuilder(\HTMLPurifier::class)
+            ->getMockBuilder(\OCA\News\Vendor\HTMLPurifier::class)
             ->disableOriginalConstructor()
+            ->onlyMethods(['purify'])
             ->getMock();
 
         $this->time = 333333;